clean up dropdown and move to separate folder. Export through index
