+++ src/core/Dropdown/index.ts
@@ -1,0 +1,2 @@
+export { Dropdown, DropdownProps } from './Dropdown/Dropdown';
+export { DropdownItem, DropdownItemProps } from './DropdownItem/DropdownItem';
